chore(cosmetic): de-uppercase form labels
Recent design proposals point towards removing the uppercase usage
across the board, and this PR starts the process with form labels.

it builds upon https://github.com/apache/superset/pull/24736 that
started to address hierarchy-is-unclear issue in Explore control panel
and makes things a bit neater on top of it.